Refactoring more Landing pages
