+++ src/LandingPages/LandingPage/LandingPage.jsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import LandingPageNav from '../components/TopNav/LandingPageNav';
-import LandingPageFooter from '../components/Footer/LandingPageFooter';
+import LandingPageNav from '../components/LandingPageHeader/LandingPageNav';
+import LandingPageFooter from '../components/LandingPageFooter/LandingPageFooter';
 import Modal from '../../components/Modal/modal';
 import './LandingPage.css';
 import { FaLinkedin } from 'react-icons/fa';
